+++ packages/eslint-plugin/tests/lib/rules/no-object-literal-type-assertion.js
@@ -39,7 +39,23 @@ ruleTester.run('no-object-literal-type-assertion', rule, {
     `const foo = <any> {};`,
     // Allow cast to 'unknown'
     `const foo = {} as unknown;`,
-    `const foo = <unknown> {};`
+    `const foo = <unknown> {};`,
+    {
+      code: `print({ bar: 5 } as Foo)`,
+      options: [
+        {
+          allowAsParameter: true
+        }
+      ]
+    },
+    {
+      code: `new print({ bar: 5 } as Foo)`,
+      options: [
+        {
+          allowAsParameter: true
+        }
+      ]
+    }
   ],
   invalid: [
     {
@@ -64,6 +80,26 @@ ruleTester.run('no-object-literal-type-assertion', rule, {
     },
     {
       code: `const x = {} as T;`,
+      errors: [
+        {
+          messageId: 'unexpectedTypeAssertion',
+          line: 1,
+          column: 11
+        }
+      ]
+    },
+    {
+      code: `print({ bar: 5 } as Foo)`,
+      errors: [
+        {
+          messageId: 'unexpectedTypeAssertion',
+          line: 1,
+          column: 7
+        }
+      ]
+    },
+    {
+      code: `new print({ bar: 5 } as Foo)`,
       errors: [
         {
           messageId: 'unexpectedTypeAssertion',